[MEMORY] Add a tria array reallocation macro.
This macro is called whenever we try to add a new tria and there is no
available slot in the tria array. It reallocates the tria array according to a
multiple of the initial size (2), as well as the adjacency table. It is quite
dirty, so it may change in the future.