[OPTIM] Add a Laplacian smoothing function.
Perform a Laplacian smoothing of internal points after each iteration of the
point insertion process in order to get better quality elements. The overall
point insertion scheme is as follow : create points to saturate the edges of
mesh N, filter them and insert them in the mesh, perform laplacian smoothing to
get mesh (N+1), go back to step 1 with the mesh (N+1).

Some topological optimizations such as edge swaps and vertex
deletions (collapses) will be implemented in the future and used in the
optimization step along with the laplacian smoothing procedure.